# openapi.yaml
openapi: 3.0.3
...
info:
  title: 夜间代客洗车平台 API
  version: 1.0.0
  description: |
    V1.0 MVP 接口契约。

    **全局约定**
    - 金额：整数，单位【分】。前端展示时除以 100。
    - 时间：毫秒时间戳（number）。时区 Asia/Shanghai。
    - 统一响应体 `{ code, msg, data }`，code = 0 表示成功，非 0 为错误码。
    - 分页响应：data = `{ list, total, pageNum, pageSize }`。
    - 幂等：下单、支付、开箱、分配格口四类接口必须传 `Idempotency-Key` 请求头。

    **路径前缀**
    - `/api/v1/**` C 端与 B 端小程序（wx-token 鉴权）
    - `/admin-api/**` 运营后台（若依 JWT 鉴权，本文件 V1.0 不展开）
    - `/device-callback/v1/**` 钥匙柜回调（签名 + IP 白名单，见 x-roadmap）
  contact:
    name: CarGlow
servers:
  - url: https://{host}
    variables:
      host:
        default: api.carglow.local
        description: 部署域名，上线后替换
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/coupons/center:
    get:
      tags: [coupon]
      summary: 领券中心
      description: 返回当前可领的券模板，含该会员已领数量与剩余库存。
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/CouponTemplateVO' }
  /api/v1/coupons/templates/{templateId}/receive:
    post:
      tags: [coupon]
      summary: 领取优惠券
      security: [{ bearerAuth: [] }]
      parameters:
        - name: templateId
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CouponUserVO' }
  /api/v1/coupons:
    get:
      tags: [coupon]
      summary: 我的优惠券
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [UNUSED, USED, EXPIRED] }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/CouponUserVO' }
  /api/v1/coupons/best:
    get:
      tags: [coupon]
      summary: 下单页算最优券
      description: 给定支付金额，返回可用且减免最多的券（可能为空）。
      security: [{ bearerAuth: [] }]
      parameters:
        - name: amount
          in: query
          required: true
          schema: { type: integer, format: int64 }
          description: 待支付金额（分）
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Result'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CouponBestVO' }

  # ============================ 保险 / 保单（C 端） ============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 小程序侧 wx-token；后台侧若依 token
    deviceSign:
      type: apiKey
      in: header
      name: X-Signature
      description: 柜机回调签名，配合 X-Device-No、X-Timestamp 与 IP 白名单
...
  schemas:
    CouponTemplateVO:
      type: object
      properties:
        templateId: { type: integer, format: int64 }
        name: { type: string }
        type: { type: string, enum: [NORMAL, NEWBIE] }
        thresholdAmount: { type: integer, format: int64, description: 使用门槛（分） }
        discountAmount: { type: integer, format: int64, description: 减免金额（分） }
        perLimit: { type: integer }
        endTime: { type: integer, format: int64, description: 失效时间（毫秒） }
        remain: { type: integer, description: 剩余库存 }
        claimed: { type: integer, description: 当前会员已领数量 }
    CouponUserVO:
      type: object
      properties:
        couponUserId: { type: integer, format: int64 }
        templateId: { type: integer, format: int64 }
        name: { type: string }
        thresholdAmount: { type: integer, format: int64 }
        discountAmount: { type: integer, format: int64 }
        status: { type: string, enum: [UNUSED, USED, EXPIRED] }
        expireTime: { type: integer, format: int64 }
        orderNo: { type: string }
    CouponBestVO:
      type: object
      properties:
        couponUserId: { type: integer, format: int64 }
        discountAmount: { type: integer, format: int64 }
...
    Result:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          description: |
            0 = 成功，非 0 见 x-error-codes。
            数字映射规则：分段字母转前缀（A=1 / B=2 / C=3）+ 4 位编号，
            即 A0002 → 10002、B1002 → 21002、C0001 → 30001。
          example: 0
        msg:
          type: string
          example: 操作成功
        data:
          description: 业务数据，各接口不同
        traceId:
          type: string
          description: 链路追踪 ID，排查问题时提供

    # ---------------- 订单状态（与 OrderStatus.java 必须逐条一致） ----------------
...